Restructure settings pages
