+++ resources/js/modules/sections/components/DeleteSectionButton.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import {t} from '@craftcms/cp/utilities/translate.ts.mjs';
-  import type {SectionModel} from '@/pages/settings/Sections.vue';
+  import type {SectionModel} from '@/pages/settings/sections/Index.vue';
   import {useForm} from '@inertiajs/vue3';
   import {destroy} from '@actions/Settings/SectionsController';
