+++ resources/js/pages/settings/assets/Index.vue
