+++ resources/js/pages/settings/entry-types/Index.vue
