+++ resources/js/pages/settings/filesystems/Index.vue
